Make update action create missing dependencies

+++ mam.py
@@ -511,7 +511,7 @@ def additional_upload(obj: str):
     json_write(f"{DIR}/objects/additionals/{obj}", {"local": version, "remote": version})
 
 
-def action_install():
+def action_setup(first: bool):
     print("Adding mam user...")
     if os.system("id mam") != 0:
         os.system("useradd --system mam")
@@ -519,16 +519,17 @@ def action_install():
     with open("/etc/sudoers.d/mam", "w") as f:
         f.write("mam ALL=(root) NOPASSWD: /usr/bin/pacman\n")
 
-    print("Installing dependencies...")
-    os.system("pacman --noconfirm -Syu git fakeroot binutils")
-    if os.system("paru --version") != 0:
-        if os.path.isdir("/tmp/paru"):
-            shutil.rmtree("/tmp/paru")
+    if first:
+        print("Installing dependencies...")
+        os.system("pacman --noconfirm -Syu git fakeroot binutils")
+        if os.system("paru --version") != 0:
+            if os.path.isdir("/tmp/paru"):
+                shutil.rmtree("/tmp/paru")
 
-        os.system("git clone https://aur.archlinux.org/paru-bin.git /tmp/paru && chown -R mam:mam /tmp/paru")
-        os.system("mkdir -p /tmp/mam && chown mam:mam /tmp/mam")
-        os.system("cd /tmp/paru && sudo -u mam HOME=/tmp/mam makepkg")
-        os.system("pacman --noconfirm -U /tmp/paru/paru-*.pkg.tar.zst")
+            os.system("git clone https://aur.archlinux.org/paru-bin.git /tmp/paru && chown -R mam:mam /tmp/paru")
+            os.system("mkdir -p /tmp/mam && chown mam:mam /tmp/mam")
+            os.system("cd /tmp/paru && sudo -u mam HOME=/tmp/mam makepkg")
+            os.system("pacman --noconfirm -U /tmp/paru/paru-*.pkg.tar.zst")
 
     print("Creating directories...")
     for type in ["directories", "files", "packages", "partials", "additionals"]:
@@ -536,8 +537,18 @@ def action_install():
         os.makedirs(f"{DIR}/backups/{type}", exist_ok=True)
 
     print("Installing mam...")
-    shutil.copy(__file__, "/usr/local/bin/mam")
-    os.chmod("/usr/local/bin/mam", 0o755)
+    if first:
+        shutil.copy(__file__, "/usr/local/bin/mam")
+        os.chmod("/usr/local/bin/mam", 0o755)
+    else:
+        req = urllib.request.Request(CONFIG["address"])
+        try:
+            res = urllib.request.urlopen(req).read().decode()
+            with open("/usr/local/bin/mam", "w") as f:
+                f.write(res)
+        except:
+            print("Could not connect to server.")
+            sys.exit(1)
 
     print("Installing daemon...")
     with open("/etc/systemd/system/mam.service", "w") as f:
@@ -557,6 +568,9 @@ def action_install():
     os.system("systemctl daemon-reload")
     os.system("systemctl enable --now mam.service")
 
+
+def action_install():
+    action_setup(True)
     print("Done!")
     print("Please run `mam auth` to authenticate this machine.")
 
@@ -629,15 +643,7 @@ def action_update():
         print("Authentication failed.")
         sys.exit(1)
 
-    req = urllib.request.Request(CONFIG["address"])
-    try:
-        res = urllib.request.urlopen(req).read().decode()
-        with open("/usr/local/bin/mam", "w") as f:
-            f.write(res)
-    except:
-        print("Could not connect to server.")
-        sys.exit(1)
-
+    action_setup(False)
     print("Updated!")
